Adds detail and more information to plotting capabilities
Add detail/information to plotting capabilities.

- Merge plot_functions.py into eval_plots.py and move eval_plots.py into the tools directory.
- Remove plots subdirectory.
- Optional argument to create barplots of CSI for each individual site.
- Create a csv containing the data used to create the scatterplots.

This resolves #325.
